requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Astro CLI documentation ---------> https://www.astronomer.io/docs/astro/cli/develop-project

psycopg2-binary
pandas
dbt-postgres
apache-airflow-providers-oracle
graphviz